+++ hack/call-sse.sh
@@ -1,18 +1,21 @@
 #!/bin/bash
 
 # A simple script to just get the content from the star-search sse.
-# Very useful when testing alongside the logs in the API.
+# Very useful when testing alongside the logs in the api.
 #
-# For testing against the live URL, use the following:
-# url="https://beta.api.opensauced.pizza/v2/star-search/stream"
+# Modify the following env vars to enable hitting the star search stream endpoint
+# against your own thread and API:
 
-url="http://localhost:3003/v2/star-search/stream"
+BEARER_TOKEN=""
+UUID=""
+URL="http://localhost:3003/v2/star-search/${UUID}/stream"
 
 # Use curl to connect to the SSE endpoint, then filter and format the output using awk
 
-curl -sN -X 'POST' "$url" \
+curl -sN -X 'POST' "${URL}" \
   -H 'accept: */*' \
   -H 'Content-Type: application/json' \
+  -H "Authorization: Bearer ${BEARER_TOKEN}" \
   -d '{
   "query_text": "What are recent pull requests in kubernetes/kubernetes?"
 }' | awk -F 'data: ' '/^data: / { printf "%s", $2 }'