+++ src/star-search/types/star-search.type.ts
@@ -1,12 +1,19 @@
 import { z } from "zod";
 
 export enum StarSearchEventTypeEnum {
+  user_prompt = "user_prompt",
   final = "final",
   content = "content",
   function_call = "function_call",
 }
 
+export enum StarSearchActorEnum {
+  user = "user",
+  manager = "manager",
+}
+
 export enum StarSearchPayloadStatusEnum {
+  recieved_user_query = "received_user_query",
   in_progress = "in_progress",
   done = "done",
 }